pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ctf-secops-platform"
version = "0.1.0"
description = "Codex-native security operations platform for CTF, KoTH, pentest, and bug bounty workflows."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.30,<3.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "psycopg[binary]>=3.1.19,<4.0.0",
  "pgvector>=0.3.2,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "mcp[cli]>=1.0.0,<2.0.0",
  "pyyaml>=6.0.1,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
]

[tool.setuptools.packages.find]
include = ["secops*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
